Add keycloak authentication for the plugin
This commit adds the keycloak authentication for the plugin. It adds the
keycloak.auth module. This module contains the KeycloakAuth struct.
This struct is responsible for authenticating the user using the
Keycloak server. An access token can be obtained from the KeycloakAuth
instance by calling the get_access_token method. This method will
return a valid access token by refreshing or re-authenticating
the token if necessary.

Adds a static lazily-initialized KeycloakAuth instance to the
nss_keycloak module (lib.rs).

Adds a integration test for the KeycloakAuth singleton in lib.rs.